Refactor duplicated credential stores and Channex sync helpers
